requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
pydantic==2.9.2
google-generativeai==0.7.2
httpx==0.27.2
python-multipart
